SimonStewart: Deleting more methods from the BCO and changing the type of the timeout to a long
r11025

+++ java/org/openqa/selenium/browserlaunchers/LauncherUtils.java
@@ -245,7 +245,7 @@ public class LauncherUtils {
 
   // TODO(simon): Reduce visibility
   public static void generatePacAndPrefJs(File customProfileDir, int port, String homePage,
-      boolean changeMaxConnections, int timeoutInSeconds, Capabilities capabilities)
+      boolean changeMaxConnections, long timeoutInSeconds, Capabilities capabilities)
       throws FileNotFoundException {
